Starting files for the dialog system
2 actor components for the NPC and player: the components will be added to the player and to any NPC the player will need to communicate to.

1 BFL (Blueprint function library) the bfl is for all of the general functions for the system and where the logic to pull the dialog data will live

several structs
Dialog sound pairs - a set of VO and the matching text line.
Character dialog grouping - set of dialog pairs and teh character name.
Dialog name line condition -[ required setup for the data table idea ] the variables can change depending on the design, but this struct is to test the possibilities of a data table hold all of the dialog

DT_CharacterLines - testing data table with character name and their lines

the Materials and textures are for the speech bubble to ID the npc that are approachable in game. and the key ID when the player approaches a NPC.

New Widgets:
W_DialogBubble - Widget component that is attached to the BP_NPC there are 2 ways of showing the NPC, either by image, text, or both.
W_dialog choice - just a textblueprint with exposed variables to update the text component from different widgets/BP
W_dialog paragraph - displays the speaker's name and their line.
W_Dialog Sselection window - until the design is confirmed. if there is a need for multiple choice for a player response, a instance of Dialogchoice widgets wiould be added into this widget depending on the orientation (vertical or horizontal).
W_Dialog main window - this holds all the log and functionality for the dialog system while in-game, whenever a player interacts with an NPC with just dialog, this window is added to the viewport.